refactor(tests): move leagueResultFixtures next to its users
It sat in utils/ documented as "for the tests here", but the test beside it
never imported it. Both consumers are in scoring/.
